+++ src/App/components/Group.js
@@ -1,0 +1,18 @@
+// Brief Discription of out Group
+import '../../css/components/Group.css';
+
+function Group() {
+  return (
+    <div className="Group">
+      <h3>Group Brian</h3>
+      <h4>Members</h4>
+      <ul className="Group-list">
+        <li>Brian</li>
+        <li>Nancy</li>
+        <li>Japheth</li>
+      </ul>
+    </div>
+  );
+}
+
+export default Group;
