+++ src/css/components/Group.css
@@ -1,0 +1,4 @@
+.Group-list {
+  padding: 0;
+  list-style-type: none;
+}
